Create AppResultsPanel component
This component contains the header icons for bar/number buttons as well
as a slot which can be used to pass in the KPI Panel. It passes the
active tab via scoped slots. This change prepares for adding more icons
that do not have a relation to the KPI Panel content