Remove scenario name from loaded project when not available
When a scenario name is removed from a workspace, we need to handle
cases where a user might have that name saved in a project.
In this case, we remove the scenarioName option from the loaded project.
This will cause the app to return to the settings page with an empty
scenario name
We also notify the user that this has happend