Ignore validation warning where Dato provides different kpi groups
Now the project validation is very tight. Meaning that if the project
kpis do not match. The project won't load. But that also means that the
server can not change without invalidating all previously saved
projects.

This PR loosens the validation. It logs to the console that the results
are not reliable. When this happens, the server should update as quickly
as possible