Load pluvfloodParam data when the project area and scenarioName are known
We need both the scenario name and the project area to fetch the
pluvflood param data. In order to make that happen, a plugin is added
which listens to both values. When they are both present, we fetch the
data.

We also check for the resulting data in the flow, that way the app
can't go ahead if the data is still loading